requirements.txt
Pillow==11.3.0
reportlab==4.4.0
google-api-python-client==2.189.0
google-auth-oauthlib==1.2.4
sounddevice==0.5.1
numpy==2.2.6
certifi==2025.7.9
PySide6==6.10.2
pyqtdarktheme==0.1.7
cryptography==46.0.4
